requirements.txt
groq>=0.4.0
faker>=20.0.0
pandas>=2.0.0
pyarrow>=14.0.0
python-docx>=1.1.0
reportlab>=4.0.0
pyyaml>=6.0
nltk>=3.8.0
spacy>=3.7.0
email-validator>=2.1.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

